+++ adventure/src/prelude.rs
@@ -5,7 +5,7 @@
 
 pub use crate::oneshot::OneshotRequest;
 pub use crate::paginator::PagedRequest;
-pub use crate::request::{BaseRequest, RepeatableRequest};
+pub use crate::request::{BaseRequest, Request};
 pub use crate::response::Response;
 pub use crate::retry::RetriableRequest;
 pub use crate::util::{RequestExt, ResponseExt};