+++ CHANGELOG.md
@@ -3,9 +3,11 @@
 
 ### Breaking Changes
 
- - `Request` is splitted into `BaseRequest` and `OneshotRequest`.
- - `RepeatableRequest` does not extend `OneshotRequest`. `OneshotRequest` and
-   `RepeatableRequest` will inherit the same base trait.
+ - `Request` is splitted into `BaseRequest` and `OneshotRequest`, and
+   `Request::into_response` is also renamed to `OneshotRequest::send_once`.
+ - `RepeatableRequest` is renamed to `Request`, and extends `BaseRequest`
+   instead `OneshotRequest`. Therefore, both of `Request` and `OneshotRequest`
+   will have the same base trait.
 
 ### New features